# Conflicts:
#	src/main/java/com/simibubi/create/Create.java
#	src/main/java/com/simibubi/create/foundation/block/CopperBlockSet.java

+++ src/main/java/com/simibubi/create/content/contraptions/fluids/potion/PotionFluid.java
@@ -33,7 +33,7 @@ public class PotionFluid extends VirtualFluid {
 		super(properties);
 	}
 
-	public static FluidStack withEffects(int amount, Potion potion, List<MobEffectInstance> customEffects) {
+	public static FluidStack withEffects(long amount, Potion potion, List<MobEffectInstance> customEffects) {
 		FluidStack fluidStack = new FluidStack(AllFluids.POTION.get()
 				.getSource(), amount);
 		addPotionToFluidStack(fluidStack, potion);
